requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.12
psycopg-binary==3.2.12
pycparser==2.23
PyJWT==2.10.1
sqlparse==0.5.3
